reorganize repo and rename it to blockstack-zones
