--- bower.json
@@ -1,20 +1,0 @@
-{
-  "name": "dns-zonefile",
-  "version": "0.1.13",
-  "homepage": "https://github.com/elgs/dns-zonefile",
-  "authors": [
-    "elgs <elgs1980@hotmail.com>"
-  ],
-  "description": "A DNS zone file parser and generator.",
-  "main": "./lib/zonefile.js",
-  "keywords": [
-    "DNS",
-    "zone"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components"
-  ]
-}
