+++ src/index.es6
@@ -1,0 +1,5 @@
+'use strict'
+
+export { makeZoneFile } from './makeZoneFile'
+export { parseZoneFile } from './parseZoneFile'
+export { ZoneFile } from './zoneFile'
